refactor: 액세스 토큰 헤더 변수명 변경에 따른 수정
